SGCRNAs

Documentation for SGCRNAs.

SGCRNAs.CGM

SGCRNAs.CorPhenMod

SGCRNAs.DrawNetwork

SGCRNAs.SetNetwork

SGCRNAs.SpectralClustering

SGCRNAs.CGM — Method

argument gene: gene name list data: gene expression matrix threshold: value for remove genes with more than a certain number of zeros mode: mode of measurement errors elimination :NONE -> measurement error is not considered (Select this option when there are sufficient number of samples; Default) :LESS -> defined as the value below the mode :SIGMA -> defined as the value below the 2σ(mode is considered as σ) :FTEST -> defined as the value below the significantly different from the mode by pval binSize: histogram bin size used to determine measurement error when mode is other than :NONE pval: p-value for determining measurement error using :FTEST & Statistical tests of correlation coefficients power: Power in statistical tests of correlation coefficients.

returns CorData: correlation matrix GradData: gradient matrix

SGCRNAs.CorPhenMod — Method

argument df1: dataframe of gene expression df2: dataframe of Phenomenon clust: cluster number of each gene (one of return value of SpectralClustering()) fn: fig save name cormode: all gene average (:AAVG) / positive correlation gene average (:PAVG) / negative correlation gene average (:NAVG) / All three types are drawn (:ALL; default) FontSize: set when drawing the average value of the correlation coefficient on the heat map

SGCRNAs.DrawNetwork — Method

fn: figure save name nw: network graph (one of return value of SetNetwork()) pos: node position (one of return value of SetNetwork()) cnctdf: converted correlation matrix (one of return value of SetNetwork()) clust: cluster number of each gene in network (one of return value of SetNetwork()) k: number of clusters nodescores: weight of node nodelabels: label of node nodescaler: multiple for node diameter adjustment (Default: 100) edgemode: mode of edges to be drawn :ALL -> All edges are drawn (Default) :N -> Only draw edges with negative values. :P -> Only draw edges with positive values. edgethreshold: Threshold value of edges to be drawn. (Default: 0.0) edgescaler: multiple for edge thickness adjustment (Default: 5) xsize, ysize: Size of the drawing area (Default: 100, 100)

SGCRNAs.SetNetwork — Method

argument df: dataframe of correlation matrix (return value of CGM()) clust: cluster number of each gene (one of return value of SpectralClustering()) pos: gene position for drawing network (one of return value of SpectralClustering()) il: module number list which you want to draw

return nw: undirected graph pos: node position cnctdf: converted correlation matrix clust: cluster number of each gene in network score: node scores

SGCRNAs.SpectralClustering — Method

argument cor: dataframe of correlation matrix (return value of CGM()) grad: dataframe of gradient matrix (return value of CGM()) tNodeNum: threshold of sub-cluster node number; default: 100 depthMax: Depth of sub-clusters; default: 5 pcas: pca dimention; default: 99 itr: number of trials; default: 300 seed: seed value of random number; default: 42 (Answer to the Ultimate Question of Life, the Universe, and Everything) nNeighbors: UMAP parameter; default: 40 minDist: UMAP parameter; default: 0.1 normFlg: Whether to symmetrically normalize the Laplacian matrix; default: false randNormFlg: Whether to random walk normalize the Laplacian matrix; default: false

returns clust: cluster number of each gene pos: gene position for drawing network edgeScore: edge score for drawing network