Reference

SGCRNAs.CGM — Function

arguments

gene::Vector: gene name list

data::Matrix: gene expression matrix

threshold::Float64: value for remove genes with more than a certain number of zeros; default: 0.5

mode::Symbol: mode of measurement errors elimination

:NONE -> measurement error is not considered (Select this option when there are sufficient number of samples; Default)

:LESS -> defined as the value below the mode

:SIGMA -> defined as the value below the 2σ(mode is considered as σ)

:FTEST -> defined as the value below the significantly different from the mode by pval

binSize::Float64: histogram bin size used to determine measurement error when mode is other than :NONE; default: 0.01

pval::Float64: p-value for determining measurement error using :FTEST & Statistical tests of correlation coefficients; default: 0.05

power::Float64: Power in statistical tests of correlation coefficients; default: 0.8

returns

CorData: correlation matrix

GradData: gradient matrix

SGCRNAs.SpectralClustering — Function

arguments

cor::DataFrame: dataframe of correlation matrix (return value of CGM())

grad::DataFrame: dataframe of gradient matrix (return value of CGM())

tNodeNum::Int64: threshold of sub-cluster node number; default: 100

depthMaxv: Depth of sub-clusters; default: 5

pcas::Int64: pca dimention; default: 99

itr::Int64: number of trials; default: 300

seed::Int64: seed value of random number; default: 42 (Answer to the Ultimate Question of Life, the Universe, and Everything)

nNeighbors::Int64: UMAP parameter; default: 40

minDist::Float64: UMAP parameter; default: 0.1

normFlg::Bool: Whether to symmetrically normalize the Laplacian matrix; default: true

randNormFlg::Bool: Whether to random walk normalize the Laplacian matrix; default: false

returns

clust: cluster number of each gene

pos: gene position for drawing network

edgeScore: edge score for drawing network

SGCRNAs.SetNetwork — Function

arguments

df::DataFrame: dataframe of correlation matrix (return value of CGM())

clust::Vector{Int64}: cluster number of each gene (one of return value of SpectralClustering())

pos::Matrix: gene position for drawing network (one of return value of SpectralClustering())

il::Vector: module number list which you want to draw

returns

nw: undirected graph

pos: node position

cnctdf: converted correlation matrix

clust: cluster number of each gene in network

score: node scores

SGCRNAs.DrawNetwork — Function

arguments

fn: figure save name

nw: network graph (one of return value of SetNetwork())

pos: node position (one of return value of SetNetwork())

cnctdf: converted correlation matrix (one of return value of SetNetwork())

clust: cluster number of each gene in network (one of return value of SetNetwork())

k: number of clusters

node_scores: weight of node

node_labels: label of node

node_scaler: multiple for node diameter adjustment; Default: 100

edge_mode: mode of edges to be drawn

:ALL -> All edges are drawn (Default)

:N -> Only draw edges with negative values

:P -> Only draw edges with positive values

edge_threshold: Threshold value of edges to be drawn; Default: 0.0

edge_scaler: multiple for edge thickness adjustment; Default: 5

xsize, ysize: Size of the drawing area; Default: 100, 100

SGCRNAs.CorPhenMod — Function

arguments

df1::DataFrame: dataframe of gene expression

df2::DataFrame: dataframe of Phenomenon

clust::Vector{Int64}: cluster number of each gene (one of return value of SpectralClustering())

fn::String: fig save name

cor_mode::Symbol: mode of caluclation of correlation coefficient

:ALL -> All three types are drawn (default)

:A_AVG -> all gene average

:P_AVG -> positive correlation gene average

:N_AVG -> negative correlation gene average